Move benchmarks to separate file
